AK: Make Vector use size_t for its size and capacity

+++ Applications/Help/History.h
@@ -38,7 +38,7 @@ public:
     void go_forward();
 
     bool can_go_back() { return m_current_history_item > 0; }
-    bool can_go_forward() { return m_current_history_item + 1 < m_items.size(); }
+    bool can_go_forward() { return m_current_history_item + 1 < static_cast<int>(m_items.size()); }
 
     void clear();
