+++ Applications/Welcome/TextWidget.cpp
@@ -62,7 +62,7 @@ void TextWidget::paint_event(GUI::PaintEvent& event)
     if (frame_thickness() > 0)
         indent = font().glyph_width('x') / 2;
 
-    for (int i = 0; i < m_lines.size(); i++) {
+    for (size_t i = 0; i < m_lines.size(); i++) {
         auto& line = m_lines[i];
 
         auto text_rect = frame_inner_rect();
