+++ DevTools/VisualBuilder/VBForm.cpp
@@ -171,7 +171,7 @@ void VBForm::keydown_event(GUI::KeyEvent& event)
             update();
             return;
         }
-        int selected_widget_index = 0;
+        size_t selected_widget_index = 0;
         for (; selected_widget_index < m_widgets.size(); ++selected_widget_index) {
             if (&m_widgets[selected_widget_index] == *m_selected_widgets.begin())
                 break;
