+++ Games/Snake/SnakeGame.cpp
@@ -66,7 +66,7 @@ void SnakeGame::reset()
 
 bool SnakeGame::is_available(const Coordinate& coord)
 {
-    for (int i = 0; i < m_tail.size(); ++i) {
+    for (size_t i = 0; i < m_tail.size(); ++i) {
         if (m_tail[i] == coord)
             return false;
     }
@@ -134,7 +134,7 @@ void SnakeGame::timer_event(Core::TimerEvent&)
 
     dirty_cells.append(m_head);
 
-    for (int i = 0; i < m_tail.size(); ++i) {
+    for (size_t i = 0; i < m_tail.size(); ++i) {
         if (m_head == m_tail[i]) {
             game_over();
             return;
